regactor [validation.js]: put validation in dedicated modul
